VOTE-3410 additional optimizations using Context
