+++ src/Context/DataProvider.jsx
@@ -1,7 +1,8 @@
 // DataProvider.js
-import { createContext } from 'react';
+import { createContext, useEffect, useState } from 'react'
 import { useFetchData } from 'Utils/useFetchData';
 import { Alert } from '@trussworks/react-uswds';
+import loadPdf from 'Utils/pdfLoader'
 
 // Data context and provider component
 const DataContext = createContext();
@@ -11,12 +12,22 @@ const lang = document.documentElement.lang;
 const locale = lang !== "en" ? `/${lang}` : '';
 
 const DataProvider = ({ children }) => {
+  const [error, setError] = useState(false);
+  const [pdfDoc, setPdfDoc] = useState(null);
+  const [form, setForm] = useState(null);
   const pageContent = useFetchData(`${locale}/nvrf/assets/pages.json`);
   const stateContent = useFetchData(`${locale}/nvrf/assets/state/${currentStateId}/data.json`, '', true);
   const stringContent = useFetchData(`${locale}/nvrf/assets/strings.json`, '6f8bb721-f017-4fcc-a826-dfc93c6759b7');
   const statesContent = useFetchData(`${locale}/nvrf/assets/states.json`);
   const cardContent = useFetchData(`${locale}/nvrf/assets/cards.json`);
   const fieldContent = useFetchData(`${locale}/nvrf/assets/fields.json`);
+
+  useEffect(() => {
+    loadPdf().then(({pdfDoc, form}) => {
+      setPdfDoc(pdfDoc);
+      setForm(form);
+    }).catch(() => setError(true));
+  }, []);
 
   let isLoading = pageContent.isLoading
     || stateContent.isLoading
@@ -29,7 +40,7 @@ const DataProvider = ({ children }) => {
     || stringContent.isError
     || statesContent.isError
     || cardContent.isError
-    || fieldContent.isError;
+    || fieldContent.isError || error;
 
   if (isLoading) return <div>Loading...</div>;
   if (isError) {
@@ -63,7 +74,7 @@ const DataProvider = ({ children }) => {
 
   return (
     <DataContext.Provider
-      value={{ pageContent, stateContent, stringContent, statesContent, cardContent, fieldContent, stepContent }}>
+      value={{ pageContent, stateContent, stringContent, statesContent, cardContent, fieldContent, stepContent, pdfDoc, form }}>
       {children}
     </DataContext.Provider>
   );
